+++ C_Game_Project/C_Game_Project/C_Game_Project/BlockerManager.h
@@ -23,7 +23,7 @@ void F_BlockerManager_Exit();
 // Utility
 //----------------------------------------------------------------------------*/
 /* Spawn: Red/Blue/Green bullet */
-void F_BlockerManager_SpawnBlocker(ObjectType type, float posX, float posY);
+void F_BlockerManager_SpawnBlocker(ObjectType type, int posX, int posY);
 
 /* Destroy: Bullet */
 void F_BlockerManager_KillBlocker(int index);
